Feat/rota atualizar produto

+++ src/main/java/com/compassuol/sp/challenge/ecommerce/web/dto/ProdutoResponseDto.java
@@ -1,11 +1,13 @@
 package com.compassuol.sp.challenge.ecommerce.web.dto;
 
+import java.math.BigDecimal;
+
 import lombok.*;
 
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor @ToString
 public class ProdutoResponseDto {
     private Long id;
     private String name;
-    private Long value;
+    private BigDecimal value;
     private String description;
 }
